backend/pyproject.toml
[tool.poetry]
name = "planned"
package-mode = false
version = "0.1.0"
description = ""
authors = ["Todd Sifleet <todd.sifleet@gmail.com>"]
readme = "README.md"
packages = [{include = "planned"}]

[tool.poetry.dependencies]
python = "^3.13"
notion-client = "2.3.0"
gcsa = "2.5.1"
uvicorn = "^0.23.0"
alembic = "^1.11.0"
fastapi = "^0.100.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
pydantic-settings = "^2.8.1"
google-auth-oauthlib = "^1.2.1"
google-auth = "^2.38.0"
google-api-python-client = "^2.166.0"
starlette-context = "^0.4.0"
itsdangerous = "^2.2.0"
pytz = "^2025.2"
aiofiles = "^25.1.0"
aiohttp = "^3.13.2"
webpush = "^1.0.6"
playwright = "^1.56.0"

[tool.poetry.group.dev.dependencies]
black = ">=23.3.0"
ruff = ">=0.0.278"
mypy = ">=1.4.1"
pytest = "^8.2.0"  
pytest-cov = ">=4.1.0"
pre-commit = ">=3.3.3"
dobles = "^4.0.2"
freezegun = "^1.5.1"
pytest-vcr = "^1.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Environment configurations
[tool.planned.env]
database_url = "postgresql://planned:straderishere@postgres:5432/planned_dev"
pythonpath = "."

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",  # line length violations (handled by black)
]

[tool.ruff.isort]
known-first-party = ["planned"]
combine-as-imports = true

[tool.ruff.per-file-ignores]
"tests/**" = ["ALL"] # Ignore all rules in the tests directory
"*_test.py" = ["ALL"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "planned"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
addopts = "--cov=planned --cov-report=term-missing --cov-report=xml"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["planned"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING",
    "raise NotImplementedError",
]